refactor(word): align single-flight config and errors
