--- src/main/java/com/linglevel/api/word/service/WordSingleFlightTimeoutException.java
@@ -1,8 +1,0 @@
-package com.linglevel.api.word.service;
-
-public class WordSingleFlightTimeoutException extends RuntimeException {
-
-    public WordSingleFlightTimeoutException(String message) {
-        super(message);
-    }
-}
